added queries to the database

+++ server/internal/entity/actorFilms.go
@@ -2,7 +2,7 @@ package entity
 
 // ActorFilms represents an actor with their associated films.
 type ActorFilms struct {
-	Id       int64  `json:"-" db:"id"`
+	Id       int    `json:"-" db:"id"`
 	Name     string `json:"name" db:"name"`
 	Sex      string `json:"sex" db:"sex"`
 	Birthday string `json:"birthday" db:"birthday"`
